Buffer

+++ app/Http/Controllers/RoleController.php
@@ -6,18 +6,20 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Redirect;
 use Spatie\Permission\Models\Permission;
 use Spatie\Permission\Models\Role;
+use function abort_unless;
+use function auth;
 use function compact;
 use function view;
 
 class RoleController extends Controller
 {
-
     /**
      * @return \Illuminate\Contracts\Foundation\Application|\Illuminate\Contracts\View\Factory|\Illuminate\View\View
      */
     public function index()
     {
-        $roles = Role::all();
+        abort_unless(auth()->user()->can('View User Role'), 403);
+        $roles = Role::paginate(10);
         return view('smartTT.role.index', compact('roles'));
     }
 
@@ -27,6 +29,7 @@ class RoleController extends Controller
      */
     public function create()
     {
+        abort_unless(auth()->user()->can('Create User Role'), 403);
         $permissions = Permission::all();
         return view('smartTT.role.create', compact('permissions'));
     }
@@ -39,6 +42,7 @@ class RoleController extends Controller
      */
     public function store(Request $request)
     {
+        abort_unless(auth()->user()->can('Create User Role'), 403);
         $request->validate([
             'name' => 'required'
         ]);
@@ -60,10 +64,9 @@ class RoleController extends Controller
      */
     public function show(Role $role)
     {
-//        TODO
-//        $users = $role->users()->with('permissions')->get();
-        $permissions = $role->permissions()->get();
-        $users = $role->users()->get();
+        abort_unless(auth()->user()->can('View User Role'), 403);
+        $permissions = $role->permissions()->paginate(5, ['*'], 'permissions');
+        $users = $role->users()->paginate(5, ['*'], 'users');
         return view('smartTT.role.show', compact('role', 'permissions', 'users'));
     }
 
@@ -74,6 +77,8 @@ class RoleController extends Controller
      */
     public function edit(Role $role)
     {
+        abort_unless(auth()->user()->can('Update User Role'), 403);
+        //TODO permission
         return view('smartTT.role.edit', compact('role'));
     }
 
@@ -85,8 +90,9 @@ class RoleController extends Controller
      */
     public function update(Request $request, Role $role)
     {
+        abort_unless(auth()->user()->can('Update User Role'), 403);
         $role->update($request->all());
-        return Redirect::back();
+        return Redirect::route('role.show', ['role' => $role->id]);
     }
 
 
@@ -97,18 +103,12 @@ class RoleController extends Controller
      */
     public function destroy(Role $role)
     {
+        abort_unless(auth()->user()->can('Delete User Role'), 403);
         if ($role->users()->count() == 0) {
             $role->delete();
             return Redirect::route('role.index');
         }
         return Redirect::back()->withErrors(['error' => 'There is User in this role! Therefore you cant delete it!']);
-    }
-
-    public function addUserToRole(Request $request)
-    {
-        $request->validate([
-            'id' => 'required'
-        ]);
     }
 
     public function attachUser(Role $role, Request $request)
